Best results yet
When sorting the tabs, tabs get moved first if they have the farthest
distance to travel. Tabs that only need to move one space get moved
last. By that point, they may already be in the correct position and in
fact don't need to move at all.

This appears to keep the tab order consistently correct. Tab groups
continue to be correct as well.