Fix handling of tabs with non http/https urls
Previously, these tabs would be left at the bottom, but if
they were already in a group, then they would mess things
up.

The new behavior is to treat these tabs as having the same
group whose name is the empty string. They are sorted as
such, but are never put into a physical tab group.

A future change can add customization as to how these tabs
should be treated: sorted at end, sorted at beginning, put into
a pre-named group, etc. But that is out of scope for this change.